feat: start script
