+++ package.json
@@ -8,6 +8,7 @@
     "gw-load": "bin.js"
   },
   "scripts": {
+    "start-with-env": "node src/index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Vasco Santos",
